# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MaintenanceCreate": {
        "properties": {
          "cost": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Cost"
          },
          "priority": {
            "default": "NORMAL",
            "enum": [
              "LOW",
              "NORMAL",
              "HIGH",
              "CRITICAL"
            ],
            "title": "Priority",
            "type": "string"
          },
          "service_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Date"
          },
          "service_type": {
            "maxLength": 100,
            "minLength": 2,
            "title": "Service Type",
            "type": "string"
          },
          "trip_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trip Id"
          },
          "vehicle_id": {
            "format": "uuid",
            "title": "Vehicle Id",
            "type": "string"
          }
        },
        "required": [
          "vehicle_id",
          "service_type",
          "cost"
        ],
        "title": "MaintenanceCreate",
        "type": "object"
      },
      "MaintenanceRead": {
        "properties": {
          "cost": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Cost",
            "type": "string"
          },
          "maintenance_id": {
            "format": "uuid",
            "title": "Maintenance Id",
            "type": "string"
          },
          "priority": {
            "title": "Priority",
            "type": "string"
          },
          "service_date": {
            "format": "date",
            "title": "Service Date",
            "type": "string"
          },
          "service_type": {
            "title": "Service Type",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "trip_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trip Id"
          },
          "vehicle_id": {
            "format": "uuid",
            "title": "Vehicle Id",
            "type": "string"
          },
          "vehicle_reg_no": {
            "title": "Vehicle Reg No",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "maintenance_id",
          "vehicle_id",
          "vehicle_reg_no",
          "trip_id",
          "service_type",
          "priority",
          "cost",
          "service_date",
          "status",
          "version"
        ],
        "title": "MaintenanceRead",
        "type": "object"
      },
      "MaintenanceTransition": {
        "properties": {
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version"
        ],
        "title": "MaintenanceTransition",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "TransitOps API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/maintenance": {
      "get": {
        "operationId": "list_maintenance_api_v1_maintenance_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MaintenanceRead"
                  },
                  "title": "Response List Maintenance Api V1 Maintenance Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Maintenance",
        "tags": [
          "maintenance"
        ]
      },
      "post": {
        "operationId": "create_maintenance_api_v1_maintenance_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Maintenance",
        "tags": [
          "maintenance"
        ]
      }
    },
    "/api/v1/maintenance/{maintenance_id}/cancel": {
      "post": {
        "operationId": "cancel_maintenance_api_v1_maintenance__maintenance_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "maintenance_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Maintenance Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceTransition"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Maintenance",
        "tags": [
          "maintenance"
        ]
      }
    },
    "/api/v1/maintenance/{maintenance_id}/close": {
      "post": {
        "operationId": "close_maintenance_api_v1_maintenance__maintenance_id__close_post",
        "parameters": [
          {
            "in": "path",
            "name": "maintenance_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Maintenance Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceTransition"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Close Maintenance",
        "tags": [
          "maintenance"
        ]
      }
    },
...
  }
}